+++ js/src/main/scala/fr/hmil/roshttp/HttpDriver.scala
@@ -1,6 +1,7 @@
 package fr.hmil.roshttp
 
 import fr.hmil.roshttp.node.Modules.{HttpModule, HttpsModule}
+import fr.hmil.roshttp.response.{HttpResponse, HttpResponseFactory}
 import monifu.concurrent.Scheduler
 
 import scala.concurrent.Future
