+++ shared/src/main/scala/fr/hmil/roshttp/DriverTrait.scala
@@ -1,5 +1,6 @@
 package fr.hmil.roshttp
 
+import fr.hmil.roshttp.response.{HttpResponse, HttpResponseFactory}
 import monifu.concurrent.Scheduler
 
 import scala.concurrent.Future
