+++ shared/src/main/scala/fr/hmil/roshttp/exceptions/HttpTimeoutException.scala
@@ -1,5 +1,6 @@
-package fr.hmil.roshttp
+package fr.hmil.roshttp.exceptions
 
+import fr.hmil.roshttp.response.HttpResponse
 
 case class HttpTimeoutException(partialResponse: Option[HttpResponse]) extends
   HttpNetworkException("Request timed out")
